+++ Thunderstore/CHANGELOG.md
@@ -1,3 +1,17 @@
+## Version 1.1.1
+
+- Fix issue where dead players wouldn't drop some items
+- Fix issue where reconnecting players would always drop all items when teleporting
+
+## Version 1.1.0
+
+- Using CSync for better network communication
+- Changed BatteryDrainPercent to an integer range from 0 to 100
+- Improving init and update performance
+- Fix issue where disconnecting players wouldn't drop some items
+- Fix incorrect default value on ResetCooldownOnOrbit
+- Fix issue where changing server host synced incorrect settings
+
 ## Version 1.0.2
 
 - Setting default config to match unmodded game behavior.
@@ -11,9 +25,9 @@
 
 Initial Release
 
-- Customize Teleporter cooldown.
-- Customize Inverse Teleporter cooldown.
 - Reset cooldowns on orbit.
+- Customize Teleporter cooldown.
 - Customizable list of items to keep on Teleport.
+- Customize Inverse Teleporter cooldown.
 - Customizable list of items to keep on Inverse Teleport.
 - Customizable battery drain on Inverse Teleport.